+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - `IMGPROXY_NETWORK` config. Allows to bind on Unix socket.
+- `IMGPROXY_CACHE_CONTROL_PASSTHROUGH` config.
 
 ### Fixed
 - Fix detection of SVG starting with a comment.